Use binder and add linux support for 7.4.0
