+++ OCCT/__init__.py
@@ -16,4 +16,4 @@
 # You should have received a copy of the GNU Lesser General Public
 # License along with this library; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
-__version__ = '7.3.0.0'
+__version__ = '7.4.0.0'
