+++ ci/conda/bld.bat
@@ -3,17 +3,17 @@ cd build
 
 cmake .. -G "Ninja" ^
     -DCMAKE_BUILD_TYPE="Release" ^
-    -DENABLE_SMESH=OFF ^
-    -DENABLE_NETGEN=OFF ^
+    -DENABLE_SMESH=ON ^
+    -DENABLE_NETGEN=ON ^
     -DENABLE_FORCE=OFF ^
     -DPTHREAD_INCLUDE_DIRS:FILEPATH="%LIBRARY_PREFIX%/include"
 if errorlevel 1 exit 1
 
-ninja
+ninja -j2
 if errorlevel 1 exit 1
 
-ninja install
+ninja install -j2
 if errorlevel 1 exit 1
 
 cd ..
-python setup.py install
+python setup.py install --prefix="%PREFIX%"
