+++ ci/conda/build.sh
@@ -1,10 +1,14 @@
 #!/usr/bin/env bash
 
+# Fail on error
+set -e
+
 declare -a CMAKE_PLATFORM_FLAGS
 if [[ ${HOST} =~ .*linux.* ]]; then
     CMAKE_PLATFORM_FLAGS+=(-DCMAKE_TOOLCHAIN_FILE="${RECIPE_DIR}/cross-linux.cmake")
 fi
 
+rm -Rf build
 mkdir build
 cd build
 
@@ -17,4 +21,4 @@ cmake .. -G "Ninja" \
 ninja install
 
 cd ..
-python setup.py install
+python setup.py install --prefix="$PREFIX"
