+++ ci/conda/meta.yaml
@@ -10,13 +10,10 @@ build:
 
 requirements:
   build:
-    - {{ compiler('cxx') }}
-    - {{ cdt('libx11-devel') }}  # [linux]
-    - {{ cdt('xorg-x11-proto-devel') }}  # [linux]
-    - {{ cdt('mesa-libgl-devel') }}  # [linux]
-    - {{ cdt('libxi-devel') }}  # [linux]
+    - {{ compiler('cxx') }} # [win]
     - cmake
     - ninja
+    - pybind11
 
   host:
     - python
@@ -26,13 +23,16 @@ requirements:
 
   run:
     - python
-    - wxpython
     - occt ==7.4.0
     - smesh ==8.3.0.3  # [win]
 
 test:
+  requires:
+    - pytest
   source_files:
     - test
+  commands:
+    - pytest test -v
 
 about:
   home: https://github.com/LaughlinResearch/pyOCCT
