--- ci/conda/run_test.bat
@@ -1,2 +1,0 @@
-cd test
-python -m unittest discover
