--- ci/conda/run_test.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-cd test
-python -m unittest discover
