+++ inc/xTypes.h
@@ -45,6 +45,10 @@ typedef struct _xcolor {
 } XColor;
 */
 
+struct _XDisplay {
+
+};
+
 struct __GLXFBConfigRec {
 
 };
