+++ test/test_smesh_modules.py
@@ -17,134 +17,42 @@
 # You should have received a copy of the GNU Lesser General Public
 # License along with this library; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
-import unittest
+import os
+import sys
+import pytest
 
 
-class TestModuleImport(unittest.TestCase):
-    """
-    Test import of each module.
-    """
-
-    def test_BaseTraceCollector(self):
-        try:
-            import OCCT.BaseTraceCollector
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_BasicsGenericDestructor(self):
-        try:
-            import OCCT.BasicsGenericDestructor
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_Driver(self):
-        try:
-            import OCCT.Driver
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_DriverDAT(self):
-        try:
-            import OCCT.DriverDAT
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_DriverGMF(self):
-        try:
-            import OCCT.DriverGMF
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_DriverSTL(self):
-        try:
-            import OCCT.DriverSTL
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_DriverUNV(self):
-        try:
-            import OCCT.DriverUNV
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_LocalTraceBufferPool(self):
-        try:
-            import OCCT.LocalTraceBufferPool
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_NETGENPlugin(self):
-        try:
-            import OCCT.NETGENPlugin
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_SMDS(self):
-        try:
-            import OCCT.SMDS
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_SMDSAbs(self):
-        try:
-            import OCCT.SMDSAbs
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_SMESH(self):
-        try:
-            import OCCT.SMESH
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_SMESHDS(self):
-        try:
-            import OCCT.SMESHDS
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_StdMeshers(self):
-        try:
-            import OCCT.StdMeshers
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
-
-    def test_Utils(self):
-        try:
-            import OCCT.Utils
-            mod_ok = True
-        except ImportError:
-            mod_ok = False
-        self.assertTrue(mod_ok)
+ALL_MODULES = [
+    'BaseTraceCollector',
+    'BasicsGenericDestructor',
+    'Driver',
+    'DriverDAT',
+    'DriverGMF',
+    'DriverSTL',
+    'DriverUNV',
+    'LocalTraceBufferPool',
+    'NETGENPlugin',
+    'SMDS',
+    'SMDSAbs',
+    'SMESH',
+    'SMESHDS',
+    'StdMeshers',
+    'Utils',
+]
 
 
-if __name__ == '__main__':
-    unittest.main()
+def is_smesh_built():
+    import OCCT
+    PY_EXT = '.pyd' if sys.platform == 'win32' else '.so'
+    smesh_path = os.path.join(OCCT.__path__[0], 'SMESH' + PY_EXT)
+    return os.path.exists(smesh_path)
+
+
+@pytest.mark.skipif(not is_smesh_built(), reason="SMESH was not built")
+@pytest.mark.parametrize('mod', ALL_MODULES)
+def test_module(mod):
+    try:
+        __import__(f'OCCT.{mod}')
+    except ImportError as e:
+        print(e)
+        raise
